Version 0.1.0
Board renders the map. Pacman can be moved on user input. Wall tiles prevent the user from further moving. Can jump from one end of the map to the other.
Ghosts are still needed. Pacman should be able to win if all coins are collected. Wall tiles should be made so that it is possible to go through if the user is a few pixels off.